refactor: remove core mcp server package
