+++ packages/core-agent/src/mcp-router.ts
