+++ packages/core-agent/src/tool-registry.ts
