+++ packages/core-agent/src/tool-snapshot.ts
