+++ packages/core-agent/tests/registry-router.test.ts
