--- packages/core-mcp-server/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@dotobokuri/core-mcp-server",
-  "version": "1.4.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sbluemin/fleet-harness.git",
-    "directory": "packages/core-mcp-server"
-  },
-  "private": true,
-  "author": "sbluemin",
-  "type": "module",
-  "files": ["dist", "AGENTS.md"],
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "scripts": {
-    "build": "tsup",
-    "test": "vitest run",
-    "typecheck": "tsc --noEmit -p tsconfig.json"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "devDependencies": {
-    "@types/node": "^24.10.0",
-    "tsup": "^8.5.0",
-    "typescript": "^6.0.2",
-    "vitest": "^4.0.13"
-  }
-}
