--- packages/core-mcp-server/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./mcp-router.js";
-export * from "./tool-registry.js";
-export * from "./tool-snapshot.js";
-export type * from "./types.js";
