--- packages/core-mcp-server/tsup.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-  entry: {
-    index: "src/index.ts"
-  },
-  format: ["esm", "cjs"],
-  dts: true,
-  sourcemap: true,
-  clean: true,
-  target: "es2022"
-});
