+++ packages/fleet-carriers/src/dispatch/taskforce.ts
@@ -6,7 +6,7 @@
 
 import type { CliType } from "@dotobokuri/core-unified-agent";
 
-import type { AgentToolCtx } from "@dotobokuri/core-mcp-server";
+import type { AgentToolCtx } from "@dotobokuri/core-agent";
 import type { CarrierJobStatus as StoredCarrierJobStatus } from "../jobs/types.js";
 import type { JobPermitAccepted } from "../jobs/lifecycle.js";
 import type { ExecResult } from "@dotobokuri/core-agent";
