+++ pnpm-workspace.yaml
@@ -2,7 +2,6 @@ packages:
   - "packages/core-agent"
   - "packages/core-unified-agent"
   - "packages/fleet-infra"
-  - "packages/core-mcp-server"
   - "packages/fleet-admiral"
   - "packages/fleet-carriers"
   - "packages/fleet-wiki"
