+++ runtime/fleet-cli/package.json
@@ -39,7 +39,6 @@
     "@dotobokuri/fleet-carriers": "workspace:*",
     "@dotobokuri/fleet-gateway": "workspace:*",
     "@dotobokuri/core-agent": "workspace:*",
-    "@dotobokuri/core-mcp-server": "workspace:*",
     "@dotobokuri/fleet-infra": "workspace:*",
     "@dotobokuri/fleet-wiki": "workspace:*",
     "@dotobokuri/fleet-wiki-ui": "workspace:*",
